+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "tw5-tickets",
-  "version": "0.10.8",
+  "version": "0.10.9",
   "description": "Search TiddlyWiki5 repository issues and pull requests",
   "main": "server.js",
   "scripts": {
