+++ public/tiddlers/dev/intro.tid
@@ -1,0 +1,25 @@
+created: 20220706151549689
+modified: 20220706160000481
+tags: server
+title: Development
+icon: $:/core/images/github
+caption: Developer Introduction
+
+<$button class="bttn" actions="<<poc2go request dev/intro.tid>>" >Intro</$button>
+<$button class="bttn" actions="<<poc2go request dev/startup.tid>>" >Startup</$button>
+<$button class="bttn" actions="<<poc2go request dev/dirs.tid>>" >Directories</$button>
+<$button class="bttn" actions="<<poc2go request dev/branching.tid>>" >Branching</$button>
+<span style="float: right;"><$button class="bttn" actions="<<poc2go fetch-tostory tickets/Search/Suggest>>">{{TiddlyWiki5 Ticket Search!!buttonText}}</$button></span>
+<hr style="opacity: .5;">
+
+!!!Developers
+
+All the source for both the server and TiddlyWiki front-end are on GitHub at [[PotOfCoffee2Go/tw5-tickets|https://github.com/PotOfCoffee2Go/tw5-tickets]]. The default TiddlyWiki delivered to the browser is 'public/search.html'. Also, the 'public/index.html' directs which TiddlyWiki is delivered by default.
+
+You can run the TW5 Ticket server on your local machine that has [[git|https://git-scm.com/]] and [[node.js|https://nodejs.org]] installed. The more difficult part is to keep the your ticket database up to date.
+
+The site is part of an experiment of building a TiddlyWiki that interacts with an [[express|http://expressjs.com/]] server. Tiddlers exist both in the TiddlyWiki .HTML file as well as the server.
+
+The idea is the server delivers tiddlers, assets, and rarely web pages. This is done using HTTP and [[socket.io|https://socket.io]] which allows a more flexible interface between server and browser.
+
+{{$:/poc2go/ui/footing}}
